New Update 26/01

+++ Project_Work_My_Telegram_bot/ClassDB/ObjectPath.cs
@@ -21,7 +21,7 @@ namespace Project_Work_My_Telegram_bot.ClassDB
        
         public int IdPath { get; set; }
         public string ObjectName { get; set; } = string.Empty; 
-        public float PathLengh { get; set; }
+        public double PathLengh { get; set; }
         public DateTime DatePath { get; set; } = DateTime.Now;
         public long? UserId { get; set; }
         public User? UserPath { get; set; }
